requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
colorama==0.4.6; platform_system == 'Windows'
django==4.2.5; python_version >= '3.8'
djangorestframework==3.14.0
easyocr==1.7.1
filelock==3.12.3; python_version >= '3.8'
fsspec==2023.9.0; python_version >= '3.8'
huggingface-hub==0.16.4; python_full_version >= '3.7.0'
idna==3.4; python_version >= '3.5'
imageio==2.31.3; python_version >= '3.8'
jinja2==3.1.2; python_version >= '3.7'
lazy-loader==0.3; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
mpmath==1.3.0
networkx==3.1; python_version >= '3.8'
ninja==1.11.1
numpy==1.25.2; python_version >= '3.9'
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76; python_version >= '3.6'
packaging==23.1; python_version >= '3.7'
pillow==10.0.0; python_version >= '3.8'
pyclipper==1.3.0.post5
python-bidi==0.4.2
pytz==2023.3.post1
pywavelets==1.4.1; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
regex==2023.8.8; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-file==1.5.1
safetensors==0.3.3
scikit-image==0.21.0; python_version >= '3.8'
scipy==1.11.2; python_version < '3.13' and python_version >= '3.9'
shapely==2.0.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
sympy==1.12; python_version >= '3.8'
tifffile==2023.8.30; python_version >= '3.9'
timm==0.9.7
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
transformers==4.33.1
typing-extensions==4.7.1; python_version >= '3.7'
tzdata==2023.3; sys_platform == 'win32'
urllib3==2.0.4; python_version >= '3.7'
validators==0.22.0
